selftests/bpf: Factor out many args tests from tracing_struct
Factor out many args tests from tracing_struct and rename some function
names to make more sense.
